ISSUE #5310 - context passes function to stretch table instead of doing it itself
