+++ frontend/src/v5/ui/routes/dashboard/projects/tickets/ticketsTable/ticketsTableContent/ticketsTableContent.component.tsx
@@ -20,36 +20,25 @@ import { useContext } from 'react';
 import { FormattedMessage } from 'react-intl';
 import { useParams } from 'react-router-dom';
 import { DashboardTicketsParams } from '@/v5/ui/routes/routes.constants';
-import _ from 'lodash';
-import { DashboardListCollapse } from '@components/dashboard/dashboardList';
-import { CircledNumber } from '@controls/circledNumber/circledNumber.styles';
-import { TicketsTableGroup } from './ticketsTableGroup/ticketsTableGroup.component';
-import {  groupTickets, NEW_TICKET_ID, NONE_OPTION, SetTicketValue, UNSET } from '../ticketsTable.helper';
 import { EmptyPageView } from '../../../../../../components/shared/emptyPageView/emptyPageView.styles';
-import { Container, ScrollableContainer, Title } from './ticketsTableContent.styles';
 import { BaseProperties, IssueProperties, SafetibaseProperties } from '@/v5/ui/routes/viewer/tickets/tickets.constants';
 import { ResizableTableContextComponent, TableColumn } from '@controls/resizableTableContext/resizableTableContext';
 import { ProjectsHooksSelectors } from '@/v5/services/selectorsHooks';
 import { Transformers, useSearchParam } from '@/v5/ui/routes/useSearchParam';
 import { Spinner } from '@controls/spinnerLoader/spinnerLoader.styles';
 import { templateAlreadyFetched } from '@/v5/store/tickets/tickets.helpers';
+import { TicketsTableResizableContent, TicketsTableResizableContentProps } from './ticketsTableResizableContent/ticketsTableResizableContent.component';
 
-type TicketsTableContentProps = {
-	setTicketValue: SetTicketValue;
-	groupBy: string
-	selectedTicketId?: string;
-};
-export const TicketsTableContent = ({ setTicketValue, selectedTicketId, groupBy }: TicketsTableContentProps) => {
+export const TicketsTableContent = (props: TicketsTableResizableContentProps) => {
 	const { filteredItems } = useContext(SearchContext);
 	const { template: templateId } = useParams<DashboardTicketsParams>();
 	const [modelsIds] = useSearchParam('models', Transformers.STRING_ARRAY);
-	
-	const showModelName = modelsIds.length > 1;
 
 	const template = ProjectsHooksSelectors.selectCurrentProjectTemplateById(templateId);
 	const { config, modules } = template;
 	const hasProperties = config?.issueProperties;
 	const hasSafetibase = modules?.some((module) => module.type === 'safetibase');
+	const showModelName = modelsIds.length > 1;
 	
 	const columns: TableColumn[] = [
 		{ name: 'id', width: 80, minWidth: 25 },
@@ -85,10 +74,6 @@ export const TicketsTableContent = ({ setTicketValue, selectedTicketId, groupBy 
 		}
 		return cols;
 	};
-	
-	const onGroupNewTicket = (groupByValue: string) => (modelId: string) => {
-		setTicketValue(modelId, NEW_TICKET_ID, (groupByValue === UNSET) ? null : groupByValue);
-	};
 
 	if (!templateAlreadyFetched(template)) {
 		return (
@@ -109,46 +94,9 @@ export const TicketsTableContent = ({ setTicketValue, selectedTicketId, groupBy 
 		);
 	}
 
-	if (groupBy === NONE_OPTION || !groupBy) {
-		return (
-			<ResizableTableContextComponent columns={columns} hiddenColumns={getHiddenColumns()} columnGap={1}>
-				<ScrollableContainer>
-					<TicketsTableGroup
-						tickets={filteredItems}
-						onNewTicket={onGroupNewTicket('')}
-						onEditTicket={setTicketValue}
-						selectedTicketId={selectedTicketId}
-					/>
-				</ScrollableContainer>
-			</ResizableTableContextComponent>
-		);
-	}
-
-	const groups = groupTickets(groupBy, filteredItems);
-
 	return (
 		<ResizableTableContextComponent columns={columns} hiddenColumns={getHiddenColumns()} columnGap={1}>
-			<Container>
-				{_.entries(groups).map(([groupName, tickets]) => (
-					<DashboardListCollapse
-						title={(
-							<>
-								<Title>{groupName}</Title>
-								<CircledNumber disabled={!tickets.length}>{tickets.length}</CircledNumber>
-							</>
-						)}
-						defaultExpanded={!!tickets.length}
-						key={groupBy + groupName + templateId + tickets}
-					>
-						<TicketsTableGroup
-							tickets={tickets}
-							onNewTicket={onGroupNewTicket(groupName)}
-							onEditTicket={setTicketValue}
-							selectedTicketId={selectedTicketId}
-						/>
-					</DashboardListCollapse>
-				))}
-			</Container>
+			<TicketsTableResizableContent {...props} />
 		</ResizableTableContextComponent>
 	);
 };
